Add setting for public tickets
